# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PalmMind RAG Backend",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/documents": {
      "post": {
        "summary": "Ingest Document",
        "operationId": "ingest_document_api_v1_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_document_api_v1_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/chat": {
      "post": {
        "summary": "Converse",
        "operationId": "converse_api_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Body_ingest_document_api_v1_documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "strategy": {
            "type": "string",
            "enum": [
              "fixed",
              "paragraph"
            ],
            "title": "Strategy",
            "default": "fixed"
          },
          "chunk_size": {
            "type": "integer",
            "maximum": 2000.0,
            "minimum": 100.0,
            "title": "Chunk Size",
            "default": 1000
          },
          "overlap": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Overlap"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_ingest_document_api_v1_documents_post"
      },
      "BookingDraft": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time"
          }
        },
        "type": "object",
        "title": "BookingDraft"
      },
      "BookingResult": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "time": {
            "type": "string",
            "format": "time",
            "title": "Time"
          },
          "booking_id": {
            "type": "string",
            "title": "Booking Id"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "status": {
            "type": "string",
            "const": "booked",
            "title": "Status",
            "default": "booked"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "date",
          "time",
          "booking_id",
          "timezone"
        ],
        "title": "BookingResult"
      },
      "ChatRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "message": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message"
          },
          "document_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Document Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "session_id",
          "request_id",
          "message"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "title": "Sources"
          },
          "booking": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BookingResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "booking_draft": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BookingDraft"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "request_id",
          "answer"
        ],
        "title": "ChatResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "strategy": {
            "type": "string",
            "enum": [
              "fixed",
              "paragraph"
            ],
            "title": "Strategy"
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count"
          },
          "status": {
            "type": "string",
            "const": "ready",
            "title": "Status",
            "default": "ready"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "filename",
          "strategy",
          "chunk_count"
        ],
        "title": "IngestResponse"
      },
      "Source": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "chunk_index": {
            "type": "integer",
            "title": "Chunk Index"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "filename",
          "chunk_index",
          "score",
          "text"
        ],
        "title": "Source"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}